requirements.txt
python-dotenv==0.19.2
openai==1.3.0
anthropic==0.5.0
pandas==1.4.4
numpy==1.23.3
requests==2.28.1 